requirements.txt
asgiref==3.3.4
certifi==2021.10.8
charset-normalizer==2.0.10
Django==3.2.4
django-leaflet==0.28.1
django-map-widgets==0.3.2
gunicorn==20.1.0
idna==3.3
psycopg2-binary==2.9.1
pytz==2021.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.8
whitenoise==6.0.0
